pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "config_manager"
version = "1.0.0"
description = "A robust configuration manager with auto-save feature"
authors = [
    {name = "Tony Xiao", email = "tony.xiao@gmail.com"}
]
requires-python = ">=3.7"
dependencies = [
    "ruamel.yaml",
    "is_debug @ file:../is_debug"
]

[project.urls]
Homepage = "https://github.com/tony-z-arthur/config_manager"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"